FIX: Simulator.agentsContainer switched from ArrayList to LinkedHashMap,...
...a map between agent's ide to the agent itself; when removing given
agents (from the List<> deadAgentsContainer), one can quickly reach
the item of the map to be removed... adding new agents may be little more
involved than before with the ArrayList, but overall the performance gain
is tremendous;

For the simulation itself, a parallelStream() was utilized (and was
super helpful in terms of the performance). With HashMap#values() the
parallelization wasn't good and was noticeably slow, but the LinkedHasMap#values()
support again a very efective paralellization (likely because it has order
and so the data can be split into data blobs).